rename and separate out CustomAIContext
