+++ actor.py
@@ -4,6 +4,7 @@ from xml.dom.minidom import Entity
 
 class Attribute:
     name = ""
+    value = ""
     parent: Entity
